% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitractive-package.R
\docType{package}
\name{knitractive-package}
\alias{knitractive}
\alias{knitractive-package}
\title{knitractive: A knitr engine to simulate interactive sessions}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This package provides a knitr engine that allows you to simulate
    interactive sessions (e.g., Python console, Bash shell) across multiple
    code chunks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datascienceworkshops/knitractive}
  \item Report bugs at \url{https://github.com/datascienceworkshops/knitractive/issues}
}

}
\author{
\strong{Maintainer}: Jeroen Janssens \email{jeroen@datascienceworkshops.com} (\href{https://orcid.org/0000-0002-5878-8924}{ORCID})

Other contributors:
\itemize{
  \item Data Science Workshops [copyright holder, funder]
}

}
\keyword{internal}
