% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_engine.R
\name{start}
\alias{start}
\title{Start a new ActiveEngine}
\usage{
start(
  name,
  command = "",
  language = name,
  prompt = NULL,
  keep_session = FALSE,
  session_width = 80,
  session_height = 24
)
}
\description{
Start a new ActiveEngine
}
