% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_engine.R
\docType{class}
\name{ActiveEngine}
\alias{ActiveEngine}
\title{Class providing ActiveEngine object.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ActiveEngine
}
\description{
Class providing ActiveEngine object.
}
\keyword{datasets}
